Remove scrolling from select visibility test (#1210)
## 🤨 Rationale

Fixes #1208 

## 👩‍💻 Implementation

Use `document` as the root element in the IntersectionObserver. This is
not currently supported on Safari, but since we do not test on Safari
(yet), it is not a problem. I've added a comment to the source about
this lack of support.

## 🧪 Testing

Test passes

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.