Fix bad import (#1215)
## 🤨 Rationale

The nimble table directive in nimble-angular now includes an import for
a file that doesn't exist. This breaks any app trying to import that
version of the package.

## 👩‍💻 Implementation

The import is for a page object under the tests directory, which does
not exist in published packages. I moved the file from the `tests`
directory to the `testing` directory, which does get included in the
published package. I updated the import (in `@ni/nimble-angular`) to
point to the new location. I also updated our nimble-angular eslint
rules to error when production code tries to import test code (which is
how we got into this situation).

## 🧪 Testing

Verified that the packed npm package contains the page object file. The
build should verify that the tests can still use the page object.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.