Add support for CTRL and SHIFT clicks in the table (#1218)
# Pull Request

## 🤨 Rationale

This is part of #856. It adds support for CTRL and SHIFT selecting in
the table.

- SHIFT + click allows selecting a range of rows. If the click ends on a
group row, everything in the group will be selected. Otherwise, group
rows are skipped within the selection range and their selection state is
calculated based on the selection state of their child rows.
- CTRL + click allows selecting/deselecting individual rows without
clearing the rest of the selection.

## 👩‍💻 Implementation

- Created `InteractiveSelectionManager` class that handles all the logic
for selecting/deselecting rows based on interactions -- clicking a row,
toggling the selection checkbox, and opening an action menu
- The `InteractiveSelectionManager` contains an instance of the abstract
`SelectionManagerBase` class so that the behavior can adjust based on
the selection mode, and the table doesn't have to swap out
implementations of its own manager.

## 🧪 Testing

- Manually tested within storybook
- Wrote some new unit tests

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>