Formatted text table columns HLD and component naming conventions (#1224)
# Pull Request

## 🤨 Rationale

Start the HLD for #1011 and other column types that display data as
text. This update focuses on shared base classes, client-specified
custom colums, and possible elements that Nimble could define, not APIs
for specific elements.

We previously brainstormed this in #1054 but I started a new PR for this
more concrete proposal.

## 👩‍💻 Implementation

This includes the HLD and also some updates to component naming
conventions which fell out of the brainstorming.

## 🧪 Testing

N/A

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: mollykreis <20542556+mollykreis@users.noreply.github.com>
Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>