HLD for event emitted when interactively changing a table column's configuration (#1240)
# Pull Request

## 🤨 Rationale

We need a plan for event(s) that get emitted when a user interactively
changes a table column's configuration, such as by sorting a column.
This PR adds an HLD describing that plan.

## 👩‍💻 Implementation

N/A

## 🧪 Testing

N/A

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.