Add column-configuration-change event to the nimble-table (#1247)
# Pull Request

## 🤨 Rationale

Some clients of the nimble-table will need to react to a user
interactively changing a column's configuration.

This PR implements the event described in #1240 

## 👩‍💻 Implementation

Add event to the web component and to Blazor.
Export event detail types from Angular.
Added event handler to storybook stories for the table.

## 🧪 Testing

- Manually tested that the event is picked up by our storybook stories
- Manually tested that the Blazor event is fired correctly
- Added a few new unit tests to nimble-components

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.