Add column configuration validation (#1246)
## 🤨 Rationale

The design for the mapping/icon columns includes a fair bit of
configuration that has constraints. We need a way for invalid
configurations to be reported in some form.

## 👩‍💻 Implementation

To the base column, we add a couple validation related functions and a
flag in the column internals. By default, the validation functions
always report a valid configuration and return an empty validation
object. Column types may define custom validation objects that verfiy
specific expectations about that column's configuration. The table's
validation object gets a new validation flag that is set when any of the
table's columns have an invalid configuration. The table watches each
column's configuration validation state, which is an observable property
of the column internals. When it changes, the table updates its flag
representing the columns' configuration validity.

## 🧪 Testing

Wrote a few unit tests to cover existing column types and the logic of
the table's flag.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.