Switch columnInternals to property initialization (#1264)
# Pull Request

## 🤨 Rationale

Prior to this change the `TableColumn.columnInternals` property was
instantiated in the `TableColumn` constructor. Coupled with the behavior
that property initialization runs before constructor execution ([see
example jsbin](https://jsbin.com/gikohuxake/edit?js,console)) this would
cause the following issue:

1. Attribute default values would be assigned in a class (such as
`false` assigned to a hidden attribute)
2. The associated observable change method would run, i.e.
`hiddenChanged`
3. If the observable change method would access `columnInternals` the
reference would be `undefined`
4. After all property initialization the constructor would run and
column Internals would be defined

This change prevents this temporal deadzone of access to
`columnInternals` by instantiating it as the first property in the base
class.

## 👩‍💻 Implementation

Instantiate the `columnInternals` property in the `TableColumn` base
class which retrieves its static configuration via the implementation of
a `getColumnInternalsOptions` method which is accessible during property
initialization.

## 🧪 Testing

Refactor relying on exisiting testing. Although was able to remove a
test that checked an error was reported if `super` was called without a
`ColumnInternalsOptions` argument.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.