Make combobox and select dropdowns float above other UI (#1244)
## 🤨 Rationale

Fixes: #606
Fixes: #1227

## 👩‍💻 Implementation

I initially looked at using [Floating-UI](https://floating-ui.com/), as
that is what FAST 2.0 is using to solve this problem. (see
https://github.com/ni/nimble/pull/1234) However, for consistency, that
would require adding support to both the Select/Combobox as well as
updating the menu button, the menu item (for submenu support), and the
tooltip. Each case was a significant change, and I ran into several
problems, including with the action menu in the table. It turned into a
40+ file change with a lot of risk. I sunk more time into it than I
wanted, and still didn't feel close to done. I'm instead taking a much
more scoped approach of extending the existing pattern and using the
anchored region in the Select and Combobox.

This required forking the templates for both controls from FAST, so that
we could add the anchored region element.

## 🧪 Testing

I added some tests to cover certain expectations about the templates,
e.g. attribute forwarding. I'm relying mostly on existing Chromatic
tests to ensure the listbox pops up in the same place given the position
options.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.