Re-enable Blazor linting (#1283)
# Pull Request

## 🤨 Rationale

#1280 

## 👩‍💻 Implementation

It looks like a the build agent began pulling in a new version of the
.NET tools. `dotnet --info` prints out this for the CI build:
```
.NET SDKs installed:
  6.0.408 [/usr/share/dotnet/sdk]
  7.0.105 [/usr/share/dotnet/sdk]
  7.0.203 [/usr/share/dotnet/sdk]
  7.0.302 [/usr/share/dotnet/sdk]
```
So I'm guessing that 7.0.302 is causing issues.

I found
[dotnet/format#1800](https://github.com/dotnet/format/issues/1800) which
is pretty close to what we're seeing. [This
comment](https://github.com/dotnet/format/issues/1800#issuecomment-1558031878)
makes it sound like there'll be a newer 7.0.3xx version that will fix
this problem, but the issue is still open so I guess that version isn't
out yet.
Also see
[dotnet/format#1834](https://github.com/dotnet/format/issues/1834) and
[dotnet/sdk#32598](https://github.com/dotnet/sdk/issues/32598) which
reference 7.0.302 and the DLL v4.6.0.0 specifically.

Changes in this PR:
- Re-enable Blazor C# linting in package.json
- Add a
[global.json](https://learn.microsoft.com/en-us/dotnet/core/tools/global-json)
file to ensure the `dotnet` commands are using .NET 6.0.xxx tooling.
- This will ensure both the CI and local builds are using the .NET 6 SDK
for the `dotnet` CLI commands
- Add a pipeline step to print out `dotnet --info`, which will print out
all the installed .NET SDK versions, which should make it easier to
debug issues like this in the future
- Update docs to explicitly call out that a .NET SDK install `6.0.202 <=
v < 7` is required
- Update the CI build (`main.yml`) `setup-dotnet` action to refer to
global.json to get the required .NET version to install (so we're not
encoding the required version in multiple places in code)

## 🧪 Testing

CI build + re-built locally.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.