Mark 'placeholder' as an @Input on the text column directive (#1286)
# Pull Request

## 🤨 Rationale

`placeholder` had a setter in the `NimbleTableColumnTextDirective`, but
it wasn't marked as an `@Input`. As a result, I cannot bind to it in an
Angular app with strict template checking enabled.

## 👩‍💻 Implementation

Mark `placeholder` as an `@Input`.

## 🧪 Testing

Manually verified that I can bind to the column's `placeholder` value
from an Angular app when I could not before.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.