Fix GroupRow component template. (#1294)
# Pull Request

## 🤨 Rationale

Noticed a mistake in the template for the `GroupRow` component, and
decided to make a quick fix for it. The existing mistake in the template
(an incorrect `span` closing tag for a `div` element) is apparently
innocuous, as FAST seems to resolve the errant tag to the same element
as was used for the start tag.