Fix broken Storybook links (#1299)
# Pull Request

## 🤨 Rationale
In #1292 I changed the titles of several stories and their URLs changed.
Various docs linked to them and those links were now broken.

Tracked by #1296. I'll leave that open for now to track the suggestion
to move the component status table to storybook.

## 👩‍💻 Implementation

Search for links to storybook and replace them with the correct links.

Update the spreadsheet that generates the component table.

## 🧪 Testing

N/A

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.