Move component status table to Storybook (#1312)
# Pull Request

## 🤨 Rationale

Milan suggested moving the status table from README to Storybook. This
makes sense because Storybook is the primary place for client-facing
documentation.

Resolves #1296 

## 👩‍💻 Implementation

1. Create a story that contains the Nimble table with columns showing
component status. This is not directly useful so I put it in a new
"Internal" section in storybook and marked it not to be part of
Chromatic tests.
2. Add an MDX file that imports that story and adds more markdown
content around it. This appears next to the Getting Started page.
3. Add MDX files to appear at the top of other sections (Tests,
Internal) and provide context about their contents.
4. Update docs that pointed at the status table.

## 🧪 Testing

Spot checked links.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Fred Visser <1458528+fredvisser@users.noreply.github.com>