Fix intermittent hang in tokens build (#1313)
# Pull Request

## 🤨 Rationale

Fixes #1297

## 👩‍💻 Implementation

The hang was happening when building two icon buffers (i.e. 16x16 and
32x32) at the same time. The solution is to inline the svg-to-ico
function (from the svg-to-ico npm package) and modify it to run
serially.

I did not seriously consider upstreaming this fix, because I'm removing
parallelism, and I don't understand the cause/breadth of this issue
enough to push that on all users of the package. The repo did not have a
single issue opened in it (ever). It has a total of 15 commits, and is
rarely touched. So the likelihood of having a change merged also seems
slim.

## 🧪 Testing

Was able to force the hang reliably by starting to delete my
node_modules folder, then cancelling when it popped up a dialog about
files being used by other processes. Doing this would always cause the
next `npm run build` of `nimble-tokens` to hang. After my changes, it no
longer hangs.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Malcolm Smith <20709258+msmithNI@users.noreply.github.com>