Initial Angular code bring up for nimble-rich-text-viewer component (#1336)
# Pull Request

## 🤨 Rationale

This PR includes the angular integration for the
`nimble-rich-text-viewer` component, and it is intended to make the
component available to the Angular application which allows us to get
started with the other features of comments dependent on this component.

## 👩‍💻 Implementation

* Added directive and module for `nimble-rich-text-viewer`. 
* Added `nimble-rich-text-viewer` to the example client app.

**Note**: We will add the input properties and attributes once the
support for the markdown parser and other functionality in the component
( https://github.com/ni/nimble/pull/1335) is completed.

## 🧪 Testing

Added basic unit tests for the directive. 

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.