Update component status table to include some components in the backlog (#1351)
# Pull Request

## 🤨 Rationale

Fred and I noticed that the picker and banner, which are queued up for
development, didn't exist in the component status table.

## 👩‍💻 Implementation

Add components to the status table with a status of ⭕️

Since we had Figma designs for them I linked to those rather than XD and
updated an existing link too. Updating the rest of the links is tracked
as part of #1263.

## 🧪 Testing

<!---
Detail the testing done to ensure this submission meets requirements. 

Include automated/manual test additions or modifications, testing done
on a local build, private CI run results, and additional testing not
covered by automatic pull request validation.
-->

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.