Add clock exclamation icon to represent "expire" (#1347)
# Pull Request

## 🤨 Rationale

New icon request. This resolves #1323 

## 👩‍💻 Implementation

Followed the contributing docs, minus the step about updating the AI
file because that wasn't provided on the issue.

The SVG that was attached to the issue had a mangled name so I'm not
certain it matches what was exported from AI. I used the name suggested
in the issue description.

I added a synonym "expire" that was suggested in the issue description.

## 🧪 Testing

Verified that the icon renders in the correct color for different
statuses.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.