Initial code bring up of nimble-rich-text-editor component (#1341)
# Pull Request

## 🤨 Rationale

This PR includes the initial bring-up of the `nimble-rich-text-editor`
component to get started with the development. The actual functionality
of the component will be created in separate PRs.

## 👩‍💻 Implementation

- Create a story for `nimble-rich-text-editor` in the storybook with the
warning message as it is in the incubating state.
- Adds the component in the component status table with ⚠️status.
- Write basic unit test cases for the component bring-up.
- Create matrix tests for the basic rendering of the component in
different themes and hidden rich text editor.

## 🧪 Testing

Added basic unit tests for the component.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: SOLITONTECH\aagash.maridas <aagash.maridas@solitontech.com>