Add recordId to TableCell, TableCellView, and delegated event details (#1348)
## 🤨 Rationale

This is a precursor to adding Angular router support to the anchor
column. We want to be able to pass a clicked link's row id as part of
the event details of the delegated click event.

## 👩‍💻 Implementation

The delegated event is fired from the TableCellView, so it needs access
to the row id. To get it there, we also need to add it to the TableCell.
The row id gets propagated through the templates.

## 🧪 Testing

Tested in prototype of Angular router support for anchor column. Added
one automated test.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: mollykreis <20542556+mollykreis@users.noreply.github.com>