Fix icon color for disabled menu items and disabled text fields (#1357)
# Pull Request

## 🤨 Rationale

Resolves #1356 
Also addresses the color of the `start` icon in a disabled text field.

## 👩‍💻 Implementation

The icons in the `start` slot of the `nimble-menu-item`,
`nimble-anchor-menu-item`, and `nimble-text-field` did not have the
correct color when disabled. I've updated the styling to set
`${iconColor.cssCustomProperty}`.

## 🧪 Testing

- Verified locally in storybook that disabled `nimble-menu-item`
elements, disabled `nimble-anchor-menu-item` elements, and disabled
`nimble-text-field` elements with icons have the correct icon color.
- Updated the matrix test for the menu to have a disabled menu item with
an icon. Previously it only had a disabled anchor menu item with an
icon.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.