Updates to contributing docs based on recent pain points (#1370)
# Pull Request

## 🤨 Rationale

Wanted to capture formal guidance for some common pieces of PR feedback
so that we have something to point to when giving that feedback.

1. process for approving chromatic diffs
2. handling invalid component state
3. reducing owner burden

## 👩‍💻 Implementation

Changes in CONTRIBUTING.md. See PR comments for additional rationale.

## 🧪 Testing

N/A

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>