Initial version of date column (#1363)
## 🤨 Rationale

Part of #1014 

## 👩‍💻 Implementation

This is the first, basic version of the column. It does not support any
configuration. It always formats the datetime in the way that will
become the default when we later add format configuration. The column
derives from `TableColumnTextBase`.
- component defined
- initial Storybook stories/documentation created
- unit tests written

## 🧪 Testing

Tested in Storybook and created unit tests.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Jesse Attas <jattasNI@users.noreply.github.com>