Blazor integration for label providers (#1366)
# Pull Request

## 🤨 Rationale

Partially addresses #1090 (a few small followup items remain in the work
item once Angular/Blazor PRs go in)

## 👩‍💻 Implementation

See the
[HLD](https://github.com/ni/nimble/tree/main/specs/labels-and-localization)
for the implementation plan for the label providers.

The Blazor integration is fairly rudimentary (just the label provider
APIs). To fully use them in a Blazor app, the client apps need to handle
localizing the strings themselves, as outlined in the Readme updates.

## 🧪 Testing

Added basic autotests. Added label providers to example app MainLayout
(matches guidance from the Readme).

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.
- Updated Readme to add Localization guidance. Also added brief guidance
on NimbleThemeProvider since it wasn't mentioned in this Readme.