Fix tabs and anchor-tabs styling to allow tabs-toolbar content to be right aligned (#1384)
# Pull Request

## 🤨 Rationale

When fixing #451 (PR #1353), the `start` slot of the `tabs` and
`anchor-tabs` were hidden, but the `grid-template-columns` styling of
each component didn't get updated to account for not having `start`
content. As a result, any usage of the `tabs-toolbar` for right-aligned
buttons broke because the right-aligned buttons became left-aligned.

## 👩‍💻 Implementation

Updated `grid-template-columns` styling of the `tabs` and `anchor-tabs`
to account for those components no longer having a supported `start`
slot.

## 🧪 Testing

- Added test case to the `tabs` and `anchor-tabs` matrix tests to show
right-aligned content.
- Manually verified that the new buttons added to the matrix tests
renders in the correct location (previously it did not)

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.