Add essential APIs into the rich-text-editor spec (#1380)
# Pull Request

## 🤨 Rationale

As part of the comments feature, it is necessary to expose the following
APIs from the `nimble-rich-text-editor` to configure and meet the
specific requirements of comments,
1. `disabled`
2. `error-visible`
3. `error-text`

## 👩‍💻 Implementation

Added the above APIs to the design section of the
`nimble-rich-text-editor` spec.

## 🧪 Testing

<!---
Detail the testing done to ensure this submission meets requirements. 

Include automated/manual test additions or modifications, testing done
on a local build, private CI run results, and additional testing not
covered by automatic pull request validation.
-->

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.