Fix keyboard navigation with hidden anchor tabs (#1397)
## 🤨 Rationale

Part of #1383
The tabs control needs to be fixed in FAST, but we have similar flawed
logic being used for the anchor tabs.


## 👩‍💻 Implementation

The code already handles disabled tabs in the keyboard navigation logic.
We just need to expand that to handle hidden tabs the same way. I
updated `isFocusableElement()` to also look for the presence of
`hidden`.

Also, to make Home/End work properly I had to replace a call to
`isDisabledElement()` with `isFocusableElement()` (which used to be
equivalent, but now are not).

## 🧪 Testing

Wrote two new test cases and tested manually in Storybook.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.