Show url of focused anchor tab at bottom of page (#1407)
## 🤨 Rationale

Part of #1401

## 👩‍💻 Implementation

- Set `delegatesFocus=true` for anchor tab so that inner anchor gets
focus
- Removed `tabindex=-1` from native anchor element, so that it can
receive keyboard focus
- No longer sharing styling between the anchor tab and the standard tab.
Moved the shared styling back under the `tab` directory, then added a
modified copy of the contents to the anchor tab stylesheet. We have to
change the `::before` and `::after` pseudoelements to be on the `a`
rather than the `:host`, because it is now the `a`'s focus state that
they depend on.

## 🧪 Testing

Tested manually in Storybook

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.