Improved rendering performance (#1426)
# Pull Request

## 🤨 Rationale

The rendering part of the pipeline gets triggered each time a transform
event update happens, and it takes a while to get the expected response,
especially with very large amounts of data.

This can be improved by removing the sorting of the data as a part of
the transform event update and the filtering of the drawn values based
on their position relative to the transformed canvas.

## 👩‍💻 Implementation

broke apart the rendering into sorting and drawing.

created a filter which checks that dies and text can be drawn on the
canvas before drawing them.

## 🧪 Testing

local benchmark tests show up to 10x increase in rendering performance
when transform events are triggered with a data load of 1M based on the
zoom level.

![before](https://github.com/ni/nimble/assets/33986780/b2ca5fc5-48c4-4877-8aa8-60d4bb57ee7a)

![after](https://github.com/ni/nimble/assets/33986780/09b0c60a-ccc3-4916-9c8a-54e40cdf891a)


## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.