Rich text viewer | Updated the default minimum height and an example string (#1420)
# Pull Request

## 🤨 Rationale

The default minimum height of the component should be one line height
and currently it creates some space in the view comments in SLE (as in
the below image).

![image](https://github.com/ni/nimble/assets/123377523/fda09244-d94e-4cfb-89ca-75693f28e4a9)


## 👩‍💻 Implementation

* Removed having the minimum height of the component and if there is no
content, the component will behave like an empty `div` element.
* Updated example string for the link as `Absolute link` which makes
more sense for the feature it actually supports.

## 🧪 Testing

Manually test and verified the change.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.