Improve cell view typing (#1431)
# Pull Request

## 🤨 Rationale

There were a few typing inconsistencies related to creating a cell view
for the nimble-table. This PR addresses those inconsistencies and
updates usages accordingly.

The problems with the cell view that are fixed in this PR are:
1. The cell view used the non-null (`!`) assertion when declaring
`columnConfig`, which isn't correct. Within `ColumnInternals`,
`columnConfig` is declared as `public columnConfig?: TColumnConfig;`, so
it is valid for the `columnConfig` to be `undefined`. Therefore, we
should make sure the cell view type reflects that possibility.
2. When a row's record didn't contain the expected fields, the cell
state would be set to `{ cellRecord: {}, columnConfig: {} }`. This leads
to both the `cellRecord` and the `columnConfig` on the cell view to be
empty objects, which shouldn't be allowed given the typing. This also
forces cell view implementations to check for scenarios that shouldn't
be allowed given the configured TypeScript types.

## 👩‍💻 Implementation

To fix the problems described above, a few type changes were made:
1. Make `cellRecord` and `columnConfig` both optional within the
`TableCellState` interface and in the base `TableCellView` class.
2. Update the code that validates a row's record to set an undefined
`TableCellState` on the cell if the record doesn't contain the expected
fields rather than creating a `TableCellState` object with `cellRecord`
and `columnConfig` both set to `{}`.
3. Update usages to ensure they are correctly handling the possibility
of an `undefined` `cellRecord` and `columnConfig`.

## 🧪 Testing

Ran existing auto tests

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.