Enable Angular combobox to read value & provide more detailed docs (#1432)
# Pull Request

## 🤨 Rationale

I was given the feedback that our existing docs for the combobox are
confusing and misleading because we only document the behavior of the
native element without any additional documentation that explains the
behavior in our supported frameworks (or even implies that the API
behavior might be different). It is also not ideal that our previous
"recommended" workflow for reading the value associated with an
`OPTION_NOT_FOUND` Angular combobox was through the `nativeElement`.
Therefore, I've added a getter for the `value` in Angular.

I expect that at least part of the documentation and implementation will
need to be updated when all the issues captured in #1303 are addressed,
but I think this change is still good to enable clients to be more
successful with the combobox before a more complete fix is in place.

Fixes https://github.com/ni/nimble/issues/682

## 👩‍💻 Implementation

- Add getter for `value` on the Angular combobox directive
- Created `.mdx` story for the comboxbox so that we can create docs with
improved formatting.

## 🧪 Testing

- Manually verified that the doc page looks correct.
- New unit tests for reading the `value` of a combobox from Angular

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.